Refactored the business logic of the state manager
