+++ audioviz/audioviz/state_manager.py
@@ -1,39 +1,25 @@
 """State management for the visualization application.
 
-The state manager handles mode transitions, timing, and event processing.
-It provides an immutable state object that drives the render loop.
+This module provides the StateStore which holds the current state,
+and the immutable VisualizationState data structure.
 """
 
 from dataclasses import dataclass
-import time
-from typing import Optional, TYPE_CHECKING
-import ctypes
-from ctypes import c_int, byref
+from typing import Optional
 from .ui import ButtonPanel, create_button_panel, Timeline, create_timeline
 from .config import ButtonType, AppConfig, UIConfig
 
-# Helper to load SDL for robust mouse polling
-try:
-    _sdl = ctypes.CDLL("/lib/x86_64-linux-gnu/libSDL2-2.0.so.0")
-    # signature: UINT32 SDL_GetMouseState(int *x, int *y)
-    _sdl.SDL_GetMouseState.argtypes = [ctypes.POINTER(c_int), ctypes.POINTER(c_int)]
-    _sdl.SDL_GetMouseState.restype = ctypes.c_uint32
-    _HAS_SDL = True
-except Exception as e:
-    print(f"Warning: Could not load SDL2 for mouse polling: {e}")
-    _HAS_SDL = False
-
 
 @dataclass(frozen=True, slots=True)
-class StateManagerConfig:
-    """Configuration for the StateManager."""
+class StateStoreConfig:
+    """Configuration for the StateStore."""
     initial_mode: str = "bars"
     width: int = 800
     height: int = 600
     total_duration: float = 0.0
     auto_switch_interval: Optional[float] = 5.0
     button_specs: tuple[tuple[str, ButtonType], ...] = ()
-    mode_order: tuple[str, ...] = ()  # Dynamic cycle order
+    mode_order: tuple[str, ...] = ()
     app_config: "AppConfig" = None
     ui_config: "UIConfig" = None 
 
@@ -119,14 +105,15 @@ class VisualizationState:
         )
 
 
-class StateManager:
+class StateStore:
     """
-    Manages state transitions based on events and time.
+    Holds the current state of the visualization.
+    Only stores state; does not contain business logic.
     """
     
-    def __init__(self, config: StateManagerConfig):
+    def __init__(self, config: StateStoreConfig):
         """
-        Initialize the State Manager.
+        Initialize the State Store.
         
         Args:
             config: Configuration object
@@ -140,142 +127,13 @@ class StateManager:
             total_duration=config.total_duration,
             auto_switch_interval=config.auto_switch_interval,
         )
-        self._button_specs = list(config.button_specs)
-        self._mode_order = config.mode_order
-        self._app_config = config.app_config
-        self._ui_config = config.ui_config
-        self._last_switch_time = time.time()
     
     @property
     def state(self) -> VisualizationState:
         """Get the current state."""
         return self._state
     
-    @property
-    def button_specs(self) -> list[tuple[str, ButtonType]]:
-        """Get the button specifications."""
-        return self._button_specs
-    
-    def update(self, events: list[tuple[str, tuple[int, ...]]]) -> VisualizationState:
-        """
-        Process events and time, returning the new state.
-        
-        Args:
-            events: List of event tuples from renderer.poll_events()
-                   Each tuple: (event_type, params_tuple)
-        
-        Returns:
-            The updated visualization state
-        """
-        # First, apply event-driven transitions
-        self._state = self._process_events(self._state, events)
-        
-        # Poll mouse directly for robust dragging
-        if _HAS_SDL:
-            self._state = self._poll_mouse(self._state)
-
-        # Then, apply time-driven transitions (if still running)
-        if self._state.is_running and self._state.auto_switch_interval is not None:
-            current_time = time.time()
-            if current_time - self._last_switch_time >= self._state.auto_switch_interval:
-                self._switch_mode()
-        
-        return self._state
-
-    def _poll_mouse(self, state: VisualizationState) -> VisualizationState:
-        """Poll SDL mouse state directly to handle dragging."""
-        x = c_int(0)
-        y = c_int(0)
-        mask = _sdl.SDL_GetMouseState(byref(x), byref(y))
-        
-        # Left button is bit 0 (value 1)
-        is_left_down = (mask & 1) != 0
-        mx, my = x.value, y.value
-        
-        new_state = state
-        
-        if is_left_down:
-            # Check if we clicked inside timeline (start drag) or are already dragging
-            if state.timeline.hit_test(mx, my) or state.is_dragging:
-                # Calculate progress
-                progress = state.timeline.get_progress_at_x(mx)
-                seek_time = progress * state.total_duration
-                # Update state: dragging=True, request seek
-                # This ensures the knob follows mouse and audio seeks continuously
-                new_state = new_state.with_time(seek_time, is_dragging=True, seek_req=seek_time)
-        else:
-            # Mouse released
-            if state.is_dragging:
-                # Stop dragging
-                 new_state = new_state.with_time(state.current_time, is_dragging=False, seek_req=None)
-        
-        return new_state
-
-    def _process_events(self, state: VisualizationState, events: list[tuple[str, tuple[int, ...]]]) -> VisualizationState:
-        """Pure-ish function to calculate next state based on events.
-        
-        Args:
-            events: List of (event_type, params) tuples
-        """
-        new_state = state
-        for event_type, params in events:
-            match event_type:
-                case "quit":
-                    return new_state.stopped()
-                
-                case "resize":
-                    if len(params) >= 2:
-                        width, height = params[0], params[1]
-                        new_state = new_state.with_size(width, height, self._button_specs, self._ui_config)
-                
-                case "keydown":
-                    if len(params) >= 1:
-                        key = params[0]
-                        # Space bar to switch modes manually
-                        if key == 32:  # SDLK_SPACE
-                            self._switch_mode()
-                            new_state = new_state.with_mode(self._state.mode)
-                        elif key == 27:  # SDLK_ESCAPE
-                            return new_state.stopped()
-                
-                case "mousedown":
-                    if len(params) >= 2:
-                        x, y = params[0], params[1]
-                        
-                        # Note: Timeline is now handled exclusively by _poll_mouse
-                        # We only check buttons here
-                        if not state.timeline.hit_test(x, y):
-                            clicked_mode = state.button_panel.hit_test(x, y)
-                            if clicked_mode:
-                                clicked_mode = clicked_mode.lower()
-                                if clicked_mode != self._state.mode:
-                                    self._state = self._state.with_mode(clicked_mode)
-                                    self._last_switch_time = time.time()
-                                    new_state = new_state.with_mode(clicked_mode)
-        
-        return new_state
-
-    def _switch_mode(self) -> None:
-        """Switch to the next visualization mode."""
-        new_mode = self._get_next_mode(self._state.mode)
-        self._state = self._state.with_mode(new_mode)
-        self._last_switch_time = time.time()
-        print(f"Switched to mode: {new_mode}")
-
-    def _get_next_mode(self, current: str) -> str:
-        """Calculate next mode using configured mode order."""
-        modes = self._mode_order
-        if not modes:
-            return current
-            
-        try:
-            # Handle aliases if necessary
-            if current == "spectrum" and "multiband" in modes:
-                 current = "multiband"
-            elif current == "multiband" and "spectrum" in modes:
-                 current = "spectrum"
-
-            idx = modes.index(current)
-            return modes[(idx + 1) % len(modes)]
-        except ValueError:
-            return modes[0] if modes else "bars"
+    @state.setter
+    def state(self, new_state: VisualizationState) -> None:
+        """Set the current state."""
+        self._state = new_state
